# NI-AM1/labs/03/src/main/java/cz/cvut/fit/niam1/rest/Country.java
package cz.cvut.fit.niam1.rest;

public class Country {

    String id;
    String name;

    public Country(String id, String name) {
        this.id = id;
        this.name = name;
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

}